libdbus-c++: added recipe for it
to fix the build, two patches where needed.

One which just adds an include of unistd.h, as the functions pipe, write and
close defined by this include, weren't defined.

The other patch enables the build to build the tools of libdbus-c++ for the
target, which, without the patch, will always be build for the host, even when
cross compiling.

Signed-off-by: Thilo Cestonaro <thilo.cestonaro@ts.fujitsu.com>
Signed-off-by: Martin Jansa <Martin.Jansa@gmail.com>